+++ src/main/java/core/Main.java
@@ -12,7 +12,10 @@ import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.xml.sax.SAXException;
-import saxHandlers.*;
+import saxHandlers.ItemCategoryHandler;
+import saxHandlers.ServiceHandler;
+import saxHandlers.UserGroupHandler;
+import saxHandlers.WorkorderDetailHandler;
 import services.UserGroupServices;
 import services.UserServices;
